refactor: create AbstractAggregationRequestIT class + fail with OpenS… (#421)
* refactor: create AbstractAggregationRequestIT class + fail with OpenSearchException details if thrown (#407)


* feature: upload IntegrationTests reports


* fix: fixed workflow file


* fix: upload reports on failure


* refactor: add more error details to tests


* feat: update github workflows with name and retention-days


* test: create index with manual mappings


* refactor: change expDate type from Date to String


* refactor: stability check


* refactor: stability check 2


* fix: change retention-days to 7 for uploaded artifacts


* refactor: change expDate type to Long


---------
