+++ .github/workflows/test-integration-unreleased.yml
@@ -55,3 +55,11 @@ jobs:
         run: |
           cd opensearch-java
           ./gradlew clean integrationTest -Dhttps=false
+
+      - name: Upload Reports
+        if: failure()
+        uses: actions/upload-artifact@v3
+        with:
+          name: test-reports
+          path: opensearch-java/java-client/build/reports/
+          retention-days: 7
