+++ .github/workflows/test-integration.yml
@@ -36,6 +36,14 @@ jobs:
       - name: Run Integration Test
         run: ./gradlew clean integrationTest
 
+      - name: Upload Reports
+        if: failure()
+        uses: actions/upload-artifact@v3
+        with:
+          name: test-reports
+          path: java-client/build/reports/
+          retention-days: 7
+
       - name: Stop Docker
         run: |
           docker-compose --project-directory .ci/opensearch down
